changed scalar flux type in MPI comms routine

+++ src/random_ray/flat_source_domain.cpp
@@ -791,7 +791,7 @@ void FlatSourceDomain::all_reduce_replicated_source_regions()
     MPI_SUM, mpi::intracomm);
 
   MPI_Allreduce(MPI_IN_PLACE, scalar_flux_new_.data(), n_source_elements_,
-    MPI_FLOAT, MPI_SUM, mpi::intracomm);
+    MPI_DOUBLE, MPI_SUM, mpi::intracomm);
 
   simulation::time_bank_sendrecv.stop();
 #endif
